test: 예상 상금을 계산할 떼 int범위를 넘어가는 경우 잘못된 값이 나오는 오류 수정

+++ src/test/java/lotto/domain/WinningStatusTest.java
@@ -19,7 +19,7 @@ class WinningStatusTest {
             winningStatus.add(prizeType);
         }
 
-        assertThat(winningStatus.getPrizeMoney(prizeType)).isEqualTo(prizeType.getPrize() * count);
+        assertThat(winningStatus.getPrizeMoney(prizeType)).isEqualTo((double)prizeType.getPrize() * count);
     }
 
     @ParameterizedTest
